requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
nltk==3.8.1
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg==3.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
razorpay==1.4.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
setuptools==75.8.0
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
